+++ Source/Blazorise/Base/BaseValidationSuccess.cs
@@ -1,0 +1,37 @@
+﻿#region Using directives
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+using Microsoft.AspNetCore.Components;
+#endregion
+
+namespace Blazorise.Base
+{
+    public abstract class BaseValidationSuccess : BaseComponent
+    {
+        #region Members
+
+        #endregion
+
+        #region Methods
+
+        protected override void RegisterClasses()
+        {
+            ClassMapper
+                .Add( () => ClassProvider.ValidationSuccess() );
+
+            base.RegisterClasses();
+        }
+
+        #endregion
+
+        #region Properties
+
+        [CascadingParameter] protected BaseValidation ParentValidation { get; set; }
+
+        [Parameter] protected RenderFragment ChildContent { get; set; }
+
+        #endregion
+    }
+}